fix(mistral): resolve onboarding 422 errors with model compatibility settings
Problem: Mistral provider onboarding failed with 422 errors due to missing
compatibility settings for model parameters.

Root Cause: Mistral API requires max_tokens field (not max_completion_tokens)
and does not support the store parameter.

Fix:
- Extract applyMistralModelCompat() into new extensions/mistral/model-compat.ts
- Apply compat patch during onboarding to set supportsStore: false and
  maxTokensField: max_tokens
- Re-export from api.ts to avoid circular imports

Closes #56546
